+++ index.html
@@ -3,15 +3,6 @@
 
 <head>
     <title>An Empty World</title>
-    <link rel="stylesheet" type="text/css" href="css/style.css">
-    <link rel="stylesheet" type="text/css" href="css/achievements.css">
-    <link rel="stylesheet" type="text/css" href="css/combat.css">
-    <link rel="stylesheet" type="text/css" href="css/dark.css">
-    <link rel="stylesheet" type="text/css" href="css/footer.css">
-    <link rel="stylesheet" type="text/css" href="css/messages.css">
-    <link rel="stylesheet" type="text/css" href="css/progress.css">
-    <link rel="stylesheet" type="text/css" href="css/progress-bar.css">
-    <link rel="stylesheet" type="text/css" href="css/store.css">
 </head>
 
 <body>
@@ -216,7 +207,5 @@
         <!-- <p>&copy; Idle Game</p> -->
     </footer>
 </body>
-
-<script type="module" src="js/main.js"></script>
 
 </html>
