+++ js/crafting.ts
