+++ js/equipment.ts
