+++ js/inventory.ts
