+++ js/settings.ts
